requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1
numpy>=1.22,<=1.23.3
pandas>=1.4,<=1.4.3
matplotlib>=3.5.2,<=3.5.3
seaborn==0.12.0
scipy>=1.9.0,<=1.9.1
statsmodels>=0.13.4,<=0.13.5
pmdarima>=2.0.1,<=2.0.2
scalecast>=0.15.11,<=0.15.12
OpenBlender>=2.11,<=2.10
